autoconfigure: use DHCP and ARP scanning to find network settings
If DHCP fails, generate ARP requests to scan for known private
class C gateway addresses.  When we find a match, scan for
an unused IP address, by generating ARP requests until we find
an IP with no ARP replies.

Once this process has succeeded, write out our network configuration
to /var/lib/epipynet/config.json and run epipynet-apply-config.

epipynet-apply-config reads /var/lib/epipynet/config.json and
uses the 'ip' command to configure the IP address on the ethernet
device and the default route.  It also generates config for dnsmasq.

We also include a tool for, upon booting, resize the root filesystem
to the full extents of the SD card on which it resides.