server: report status via a Unix domain socket
Bind a Unix domain socket to '/var/run/epipynet/epipynet.sock' and
respond to status requests received via that socket.

When responding to a 'status' command, if there is another DHCP
server on the local network, report that device.  If multiple
status requests are made in parallel, only do the DHCP detection
a single time, and use the result from that single process
for all requests.

Periodically scan for DHCP servers.  When reporting status, if the
previous scan found no DHCP server, then immediately report that
to speed up status report time.  If there was one found, however,
then rescan to see if it has been disabled.