autoconfigure: use configuration from previous run at start
When autoconfiguring the network, apply the network configuration
from a previous run before we start probing for DHCP and
gateways.  If DHCP probes fail, first try probing for our previous
gateway, and if it is found, continue using that network config.

When a DNS server is provided in a DHCP offer, record that
server in the config file.

The tuple used to store network configuration made the code hard to
maintain, and didn't support a list of DNS servers.  Instead
use a epipynet.netconfig.NetworkConfig object to store the
network configuration information.

When in test mode, don't stop dnsmasq at startup.  We don't
want to interfere with a running system.

Also, bind unix domain socket to a test path for the socket,
also to avoid interfering with the real autoconfigure process.