dhcp-detect: report server address from UDP packet, not DHCP field
Previously, we would report the address of a conflicting DHCP server
on the network using the "server" field of the DHCP offer packet.
Some home routers fail to fill out this field with their own IP address.

Instead, we'll use the IP address from which the DHCP offer packet is
received, as reported by the socket's 'recvfrom' result.  This will
work for obtaining the DHCP server's address, regardless of the content
of the server address field in the DHCP offer.